jest.config.js - add note of location in root dir
Hello,

When setting up a custom project without create react app, I ran into an issue where the jest.config.js file was not being read by the compiler. I solved it by moving it to the root directory (instead of the __tests__ file). I realize that this may be an innocuous change, but I struggled mightily with this issue for days when trying to set up MSW. I thought the small note about this on line 92 would be helpful for others new to testing, especially with jest and MSW. 

Thanks,

David