Included ES6 Alternative set up
Love MSW and the ideas behind it. The community seems pretty cool to 😎.

I believe my dev experience with MSW would have been better had something like this been included in the docs.

The set up described before this change caused errors for me. Nesting the import(path/to/browser.ts) did not behave in the same way that a require would, and `import * as name from` wasn't possible in an if condition.

While this change might not be crucial, I really think it will be helpful to new users coming to MSW.

Hope you can consider it. Happy to edit it if there are layout issues.